src/requirements.txt
async_pymongo==0.1.4
pydantic==2.7.0
pydantic-settings==2.2.1
python-dateutil==2.9.0